feat(6784): extract kolme runtime commit client tests
