+++ crates/kamn-core/tests/kolme_runtime_commit_client/live_provider_contract_tests.rs
@@ -1,0 +1,10 @@
+pub(super) use super::support::*;
+
+#[path = "live_provider_contract_tests/constructor_validation_contract_tests.rs"]
+mod constructor_validation_contract_tests;
+#[path = "live_provider_contract_tests/malformed_response_contract_tests.rs"]
+mod malformed_response_contract_tests;
+#[path = "live_provider_contract_tests/outcome_and_finality_contract_tests.rs"]
+mod outcome_and_finality_contract_tests;
+#[path = "live_provider_contract_tests/response_mapping_contract_tests.rs"]
+mod response_mapping_contract_tests;
