+++ crates/kamn-core/tests/kolme_runtime_commit_client/request_validation_contract_tests.rs
@@ -1,0 +1,10 @@
+pub(super) use super::support::*;
+
+#[path = "request_validation_contract_tests/basic_request_contract_tests.rs"]
+mod basic_request_contract_tests;
+#[path = "request_validation_contract_tests/fail_closed_request_contract_tests.rs"]
+mod fail_closed_request_contract_tests;
+#[path = "request_validation_contract_tests/fixture_validation_contract_tests.rs"]
+mod fixture_validation_contract_tests;
+#[path = "request_validation_contract_tests/performance_request_contract_tests.rs"]
+mod performance_request_contract_tests;
